serverless-main/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.13.1
aniso8601==9.0.1
asyncio==3.4.3
attrs==23.2.0
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudevents==1.10.1
cryptography==42.0.2
DateTime==5.4
deprecation==2.1.0
distlib==0.3.8
disutils==1.4.32.post2
filelock==3.13.1
Flask==3.0.2
Flask-HTTPAuth==4.8.0
Flask-Migrate==4.0.7
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
flask-swagger-ui==4.11.1
frozenlist==1.4.1
functions-framework==3.5.0
gunicorn==21.2.0
idna==3.6
importlib-resources==6.1.1
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
jsonify==0.5
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
Mako==1.3.2
MarkupSafe==2.1.4
multidict==6.0.5
mysql-connector-python==8.3.0
packaging==23.2
passlib==1.7.4
platformdirs==4.2.0
pluggy==1.4.0
pycparser==2.21
PyMySQL==1.1.0
pytest==8.0.0
python-dotenv==1.0.1
pytz==2024.1
referencing==0.33.0
requests==2.31.0
rpds-py==0.17.1
setuptools==69.0.3
SQLAlchemy==2.0.25
SQLAlchemy-Utils==0.41.2
typing_extensions==4.9.0
urllib3==2.2.1
virtualenv==20.25.1
watchdog==4.0.0
Werkzeug==3.0.1
yarl==1.9.4
zope.interface==6.1